requirements.txt
playwright>=1.34.0
pandas>=1.3.0
openpyxl>=3.0.0
pillow>=8.0.0 